requirements_flask.txt
flask>=2.3.0
torch>=2.0.0
torchvision>=0.17.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
werkzeug>=2.3.0
easyfsl>=1.5.0
opencv-python>=4.8.0
